requirements.txt
Django==6.0.3
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2026.1
bcrypt==5.0.0
python-dotenv==1.2.2
firebase-admin==6.6.0
requests==2.32.3
whitenoise==6.9.0
gunicorn==23.0.0